Digitizer: implement acquisition start synchronization
Introduce new parameter `bridge` describing connection to the VME
bridge. If present, assume that bridge OUT0 is connected to digitizer 0
S-IN, digitizer 0 TRG-OUT is connected to digitizer 1 S-IN, and so on.
Configure the bridge to emit a pulse in `start_acquisition` and
configure digitizers to listen to the pulse and propagate it through the
chain.