Digitizer, HVoltage: detect link based on the bridge firmware
We have two VME crates with bridges connected to the readout unit. Their
USB link numbers will be assigned in ther order of their power up. With
this patch the tools will determine the correct connection by checking
the bridge firmware version. The better way to do it is to check the BA
register which value is assigned by rotary switches on the boards, but
currently they seem to be set to the same value.